Finished adding reviews section and testing for drinks

+++ server.js
@@ -90,12 +90,24 @@ app.get('/get-latest', (req, res) => {
 });
 
 app.delete('/delete/:id', (req, res) => {
-  const { id } = req.params;
+  // const { id } = req.params;
+  const id = parseInt(req.params.id, 10);  // Parse the id to an integer
+  if (isNaN(id)) {
+    return res.status(400).json({ message: 'Invalid ID format' });
+  }
   const query = 'DELETE FROM Drinks WHERE id = ?';
 
   db.getCon().query(query, [id], (error, results) => {
-    if (error) throw error;
-    res.send("Item deleted successfully");
+    // if (error) throw error;
+    // res.send("Item deleted successfully");
+    if (error) {
+      console.error('Error deleting item:', error);
+      return res.status(500).json({ message: 'Internal server error' });
+    }
+    if (results.affectedRows === 0) {
+      return res.status(404).json({ message: 'Item not found' });
+    }
+    res.status(200).json({ message: 'Item deleted successfully' });
   });
 });
 
@@ -115,15 +127,15 @@ app.post('/add-drink', upload.single('add-new-item-image'), (req, res) => {
       res.send(SUCCESS_MSG);
     });
   } else {
-    const query = 'INSERT INTO Drinks (name, description, price, type) VALUES (?, ?, ?, ?)'
+    const query = 'INSERT INTO Drinks (name, description, price, type, url) VALUES (?, ?, ?, ?, ?)'
 
-    db.getCon().query(query, [name, description, price, drinkType], (error, result) => {
+    db.getCon().query(query, [name, description, price, drinkType, url], (error, result) => {
       if (error) throw error;
       res.send(SUCCESS_MSG);
     });
   }
 
-  console.log(name, description, price, drinkType, image);
+  console.log(name, description, price, drinkType, image, url);
 
 })
 
@@ -368,6 +380,8 @@ app.post('/submit-review', async (req, res) => {
   });
 });
 
+module.exports = app;
+
 app.listen(port, () => {
   console.log(`Server is running on port ${port}`);
 });
@@ -399,4 +413,24 @@ app.post('/tokenTest',(req, res) => {
 
 app.get('/tokenTest',[authJwt.verifyToken,authJwt.verifyAdmin],(req, res) => {
   res.json("authoriztion worked")
+})
+
+app.get('/reviews', (req, res) => {
+  const page = parseInt(req.query.page) || 1;
+  const pageSize = parseInt(req.query.pageSize) || 10;
+  const offset = (page - 1) * pageSize;
+  const sql = "SELECT id, review_text FROM Reviews LIMIT ? OFFSET ?";
+  db.getCon().query(sql, [pageSize, offset], (error, results) => {
+    if (error) throw error;
+    res.json(results);
+  });
+});
+
+app.get('/get-reviews', (req, res) => {
+  const sql = "SELECT * FROM Reviews";
+  db.getCon().query(sql, (error, results) => {
+    if (error) throw error;
+    res.json(results);
+  });
+
 })
